Edit CSS in exercise 1 of foundations-cascade
Edit CSS in 01-cascade-fix exercise to get targeted image. First
the .small-para selector was edited to .para.small-para to add more
specificity with more class selectors and to specifically target that
element. Changed .confirm selector to .button.confirm for the same
reason. As well as edited .child to .text.child. for the same reason.